combine and clean up the two GenBank tests
